Updated jets and README

+++ src/com/skilldistillery/jet/CargoPlane.java
@@ -15,7 +15,7 @@ public class CargoPlane extends Jet implements Cargo{
 		super();
 	}
 
-	public CargoPlane(String model, double speed, int range, double price) {
+	public CargoPlane(String model, double speed, int range, long price) {
 		super(model, speed, range, price);
 		
 	}
