+++ src/com/skilldistillery/jet/JetImpl.java
@@ -10,7 +10,7 @@ public class JetImpl extends Jet {
 
 		super();
 	}
-	public JetImpl(String model, double speed, int range, double price) {
+	public JetImpl(String model, double speed, int range, long price) {
 		super(model, speed, range, price);
 	}
